Connect_2_PBIX.py
Identify the open port and database (also known as the 'Initial Catalog') of the open PBIX file. 
Establish a connection to the PBIX file.
Execute a DMV (Dynamic Management View) to create a list of  tables in the data model.
Iterate through the table list, loading each into a separate dataframe.